Move detecting phase rows to function
